% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install Dependencies}
\usage{
install_dependencies()
}
\description{
This function installs all necessary CRAN, Bioconductor, and GitHub packages.
}
