% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ASV.R, R/test_biomarkars.R,
%   R/test_phylo.R, R/test_picrust.R, R/test_stat.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Title}
\usage{
install_dependencies()

install_dependencies()

install_dependencies()

install_dependencies()

install_dependencies()
}
\description{
Title

Title

Title

Title

Title
}
